+++ src/graphqlUploadKoa.mjs
@@ -1,14 +1,16 @@
-import { processRequest } from './processRequest'
+import { processRequest as defaultProcessRequest } from './processRequest'
 
 /**
- * Creates [Koa](https://koajs.com) middleware that processes GraphQL multipart
- * requests using [`processRequest`]{@link processRequest}, ignoring
- * non-multipart requests. It sets the request body to be [similar to a
- * conventional GraphQL POST request]{@link GraphQLOperation} for following
- * GraphQL middleware to consume.
+ * Creates [Koa](https://koajs.com) middleware that processes
+ * [GraphQL multipart requests](https://github.com/jaydenseric/graphql-multipart-request-spec)
+ * using [`processRequest`]{@link processRequest}, ignoring non-multipart
+ * requests. It sets the request body to be
+ * [similar to a conventional GraphQL POST request]{@link GraphQLOperation} for
+ * following GraphQL middleware to consume.
  * @kind function
  * @name graphqlUploadKoa
- * @param {UploadOptions} options GraphQL upload options.
+ * @param {ProcessRequestOptions} options Middleware options. Any [`ProcessRequestOptions`]{@link ProcessRequestOptions} can be used.
+ * @param {ProcessRequestFunction} [options.processRequest=processRequest] Used to process [GraphQL multipart requests](https://github.com/jaydenseric/graphql-multipart-request-spec).
  * @returns {Function} Koa middleware.
  * @example <caption>Basic [`graphql-api-koa`](https://npm.im/graphql-api-koa) setup.</caption>
  * ```js
@@ -26,13 +28,20 @@ import { processRequest } from './processRequest'
  *   .listen(3000)
  * ```
  */
-export const graphqlUploadKoa = options => async (ctx, next) => {
+export const graphqlUploadKoa = ({
+  processRequest = defaultProcessRequest,
+  ...processRequestOptions
+} = {}) => async (ctx, next) => {
   if (!ctx.request.is('multipart/form-data')) return next()
 
   const finished = new Promise(resolve => ctx.req.on('end', resolve))
 
   try {
-    ctx.request.body = await processRequest(ctx.req, ctx.res, options)
+    ctx.request.body = await processRequest(
+      ctx.req,
+      ctx.res,
+      processRequestOptions
+    )
     await next()
   } finally {
     await finished