+++ src/index.mjs
@@ -29,10 +29,21 @@ export { graphqlUploadExpress } from './graphqlUploadExpress'
  */
 
 /**
- * GraphQL upload server options, mostly relating to security, performance and
- * limits.
+ * Processes a [GraphQL multipart request](https://github.com/jaydenseric/graphql-multipart-request-spec).
  * @kind typedef
- * @name UploadOptions
+ * @name ProcessRequestFunction
+ * @type {Function}
+ * @param {IncomingMessage} request [Node.js HTTP server request instance](https://nodejs.org/api/http.html#http_class_http_incomingmessage).
+ * @param {ServerResponse} response [Node.js HTTP server response instance](https://nodejs.org/api/http.html#http_class_http_serverresponse).
+ * @param {ProcessRequestOptions} [options] Options for processing the request.
+ * @returns {Promise<GraphQLOperation | Array<GraphQLOperation>>} GraphQL operation or batch of operations for a GraphQL server to consume (usually as the request body).
+ */
+
+/**
+ * Options for processing a [GraphQL multipart request](https://github.com/jaydenseric/graphql-multipart-request-spec);
+ * mostly relating to security, performance and limits.
+ * @kind typedef
+ * @name ProcessRequestOptions
  * @type {object}
  * @prop {number} [maxFieldSize=1000000] Maximum allowed non-file multipart form field size in bytes; enough for your queries.
  * @prop {number} [maxFileSize=Infinity] Maximum allowed file size in bytes.